refactor(donor-research): move body validation out of the preview component
getOutreachBodyIssue lives in its own module so the component file only
exports components (Fast Refresh), and the standalone issue type union
is inlined into the signature.
